Get rid of the systemArraycopy helper by making it an intrinsic.
Intrinsics are a bit like primitives, except that they are special
for the *optimizer*, not for the compiler. This allows to defer
their optimization to link time, if a correct (but non-optimal)
implementation can be written in Scala.js.

This is the case for System.arraycopy, which is the first
primitive to be turned into an intrinsic.

We call these things intrinsics because they are very similar
to JVM intrinsics: methods with a correct implementation in either
pure Java or through JNI, but that the optimizing compilers know
how to handle better.