Make java.lang.Character inlineable.
This allows box/unbox pairs of Chars to be eliminated.

We add an exception saying that an inlineable Character should not
cause a method to be inlined just because we pass it the boxed Char.
We only want *local* box/unbox pairs to be eliminated, similarly to
"box"/unbox pairs of other primitive types.