Fix #1718: Validate the regex directly in Pattern.compile().
Since this requires the creation of a js.RegExp at the time
Pattern.compile() is called anyway, we also store that native
RegExp in the Pattern instance. RegExp's for the Matchers are
created from that model RegExp, instead of from pattern and
flags every time. This *might* be optimized in JS VMs to reuse
the internally compiled representation for the regex.

A bug in PhantomJS 1.x causes new RegExp(r) to return the same
object r rather than a new one. So we work around that bug by
reconstructing the arguments necessary to create a new RegExp
if we detect that the same object was returned.